requirements.txt
argparse==1.2.1
Django==1.5.4
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
google-api-python-client==1.2
gunicorn==18.0
oauthlib==0.6.0
psycopg2==2.5.1
python-openid==2.2.5
python-social-auth==0.1.13
requests==2.0.0
requests-oauthlib==0.3.2
six==1.4.1
South==0.8.2
static==0.4
twython==3.1.0
wsgiref==0.1.2
Mock==1.0.1
factory_boy==2.2.1
